UFO/requirements.txt
art==6.1
colorama==0.4.6
langchain==0.3.24
langchain_community==0.3.22
msal==1.31.0
openai==1.66.2
Pillow==10.4.0
pywin32==310
pywinauto==0.6.8
PyYAML==6.0.1
requests==2.32.3
faiss-cpu==1.8.0
numpy>=2.1.0
lxml==5.1.0
psutil==5.9.8
beautifulsoup4==4.12.3
sentence-transformers==2.6.0
langchain-huggingface==0.1.2
pandas==2.0.3
html2text==2024.2.26
pyautogui==0.9.54
uiautomation==2.0.18
gradio-client==1.7.1
##For removing stopwords
#nltk==3.8.1
##For Gemini
# google-genai==1.12.1


## If use AAD to authenticate
azure_identity==1.20.0
azure-identity-broker==1.2.0